+++ pos/pkg/services/response.go
@@ -4,14 +4,14 @@ import "github.com/maslow123/pos/pkg/pb"
 
 func genericCreatePosResponse(statusCode int, errorMessage string) (*pb.CreatePosResponse, error) {
 	return &pb.CreatePosResponse{
-		Status: int64(statusCode),
+		Status: int32(statusCode),
 		Error:  errorMessage,
 	}, nil
 }
 
 func genericPosDetailResponse(statusCode int, errorMessage string) (*pb.PosDetailResponse, error) {
 	return &pb.PosDetailResponse{
-		Status: int64(statusCode),
+		Status: int32(statusCode),
 		Error:  errorMessage,
 	}, nil
 }
@@ -25,21 +25,21 @@ func genericListPosByUserResponse(statusCode int, errorMessage string) (*pb.GetP
 
 func genericUpdatePosByUserResponse(statusCode int, errorMessage string) (*pb.UpdatePosResponse, error) {
 	return &pb.UpdatePosResponse{
-		Status: int64(statusCode),
+		Status: int32(statusCode),
 		Error:  errorMessage,
 	}, nil
 }
 
 func genericDeletePosByUserResponse(statusCode int, errorMessage string) (*pb.DeletePosResponse, error) {
 	return &pb.DeletePosResponse{
-		Status: int64(statusCode),
+		Status: int32(statusCode),
 		Error:  errorMessage,
 	}, nil
 }
 
 func genericUpdateTotalPosByUserResponse(statusCode int, errorMessage string) (*pb.UpdateTotalPosResponse, error) {
 	return &pb.UpdateTotalPosResponse{
-		Status: int64(statusCode),
+		Status: int32(statusCode),
 		Error:  errorMessage,
 	}, nil
 }
